extract Theme component from MasterLayout
